+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Return user from userState `set` function
 * Refactor the placement engine to use promises
 * Update `placementEngine` function to take `promos`, `storage`, and `userAgent` arguments
+* Fix bug in `xeqBy` function
 
 ## 3.0.0 (2019-04-30)